hamoshedit

+++ Src/Core/MarketZone.Application/Features/Sales/Commands/DeleteSalesInvoice/DeleteSalesInvoiceCommandHandler.cs
@@ -4,6 +4,7 @@ using MarketZone.Application.Helpers;
 using MarketZone.Application.Interfaces;
 using MarketZone.Application.Interfaces.Repositories;
 using MarketZone.Application.Wrappers;
+using MarketZone.Domain.Sales.Enums;
 
 namespace MarketZone.Application.Features.Sales.Commands.DeleteSalesInvoice
 {
@@ -15,6 +16,10 @@ namespace MarketZone.Application.Features.Sales.Commands.DeleteSalesInvoice
 			if (entity is null)
 			{
 				return new Error(ErrorCode.NotFound, translator.GetString(TranslatorMessages.SalesInvoiceMessages.SalesInvoice_NotFound_with_id(request.Id)), nameof(request.Id));
+			}
+			if (entity.Status == SalesInvoiceStatus.Posted)
+			{
+				return new Error(ErrorCode.AccessDenied, translator.GetString("SalesInvoice_Delete_NotAllowed_After_Post"), nameof(request.Id));
 			}
 			repository.Delete(entity);
 			await unitOfWork.SaveChangesAsync();
